+++ src/engine/Graphics/Animation.ts
@@ -36,14 +36,35 @@ export enum AnimationStrategy {
   Freeze = 'freeze'
 }
 
+/**
+ * Frame of animation
+ */
 export interface Frame {
+  /**
+   * Optionally specify a graphic to show, no graphic shows an empty frame
+   */
   graphic?: Graphic;
-  duration?: number; // number of ms the frame should be visible, overrides the animation duration
+  /**
+   * Optionally specify the number of ms the frame should be visible, overrids the animation duration (default 100 ms)
+   */
+  duration?: number;
 }
 
+/**
+ * Animation options for building an animation via constructor.
+ */
 export interface AnimationOptions {
+  /**
+   * List of frames in the order you wish to play them
+   */
   frames: Frame[];
+  /**
+   * Optionally specify a default frame duration in ms (Default is 1000)
+   */
   frameDuration?: number;
+  /**
+   * Optionally specify the [[AnimationStrategy]] for the Animation
+   */
   strategy?: AnimationStrategy;
 }
 
@@ -54,6 +75,11 @@ export type AnimationEvents = {
   ended: Animation;
 };
 
+/**
+ * Create an Animation given a list of [[Frames]] in [[AnimationOptions]]
+ *
+ * To create an Animation from a [[SpriteSheet]], use [[Animation.fromSpriteSheet]]
+ */
 export class Animation extends Graphic implements HasTick {
   private static _LOGGER = Logger.getInstance();
   public events = new EventDispatcher<any>(this); // TODO replace with new Emitter
@@ -89,10 +115,34 @@ export class Animation extends Graphic implements HasTick {
     });
   }
 
+  /**
+   * Create an Animation from a [[SpriteSheet]], a list of indices into the sprite sheet, a duration per frame
+   * and optional [[AnimationStrategy]]
+   *
+   * Example:
+   * ```typescript
+   * const spriteSheet = SpriteSheet.fromImageSource({
+   *   image: imageSource,
+   *   grid: {
+   *     rows: 5,
+   *     columns: 2,
+   *     spriteWidth: 32, // pixels
+   *     spriteHeight: 32, // pixels
+   *   }
+   * });
+   *
+   * const anim = Animation.fromSpriteSheet(spriteSheet, range(0, 5), 200, AnimationStrategy.Loop);
+   * ```
+   *
+   * @param spriteSheet
+   * @param frameIndices
+   * @param durationPerFrameMs
+   * @param strategy
+   */
   public static fromSpriteSheet(
     spriteSheet: SpriteSheet,
     frameIndices: number[],
-    duration: number,
+    durationPerFrameMs: number,
     strategy: AnimationStrategy = AnimationStrategy.Loop
   ): Animation {
     const maxIndex = spriteSheet.sprites.length - 1;
@@ -107,7 +157,7 @@ export class Animation extends Graphic implements HasTick {
         .filter((_, index) => frameIndices.indexOf(index) > -1)
         .map((f) => ({
           graphic: f,
-          duration: duration
+          duration: durationPerFrameMs
         })),
       strategy: strategy
     });