+++ src/spec/GraphicsSpriteSheetSpec.ts
@@ -35,7 +35,7 @@ describe('A SpriteSheet for Graphics', () => {
     const image = new ex.ImageSource('src/spec/images/GraphicsTextSpec/spritefont.png');
 
     await image.load();
-    const ss = ex.SpriteSheet.fromGrid({
+    const ss = ex.SpriteSheet.fromImageSource({
       image,
       grid: {
         rows: 3,
@@ -55,7 +55,7 @@ describe('A SpriteSheet for Graphics', () => {
 
     await image.load();
 
-    const ss = ex.SpriteSheet.fromGrid({
+    const ss = ex.SpriteSheet.fromImageSource({
       image,
       grid: {
         rows: 4,
@@ -79,6 +79,43 @@ describe('A SpriteSheet for Graphics', () => {
     expect(ss.sprites.length).toBe(4 * 14);
 
     await expectAsync(canvasElement).toEqualImage('src/spec/images/SpriteSheetSpec/NewSpriteSpacing.png');
+  });
+
+  it('can retrieve a sprite by x,y', async () => {
+    const image = new ex.ImageSource('src/spec/images/SpriteSheetSpec/kenny-cards.png');
+
+    await image.load();
+
+    const ss = ex.SpriteSheet.fromImageSource({
+      image,
+      grid: {
+        rows: 4,
+        columns: 14,
+        spriteWidth: 42,
+        spriteHeight: 60
+      },
+      spacing: {
+        originOffset: { x: 11, y: 2 },
+        margin: { x: 23, y: 5 }
+      }
+    });
+    const logger = ex.Logger.getInstance();
+    spyOn(logger, 'warn');
+
+    expect(ss.getSprite(0, 0)).withContext('top left sprite').toEqual(ss.sprites[0]);
+    expect(ss.getSprite(13, 3)).withContext('bottom right sprite').not.toBeNull();
+
+    expect(ss.getSprite(13, 4)).toBeNull();
+    expect(logger.warn).toHaveBeenCalledWith('No sprite exists in the SpriteSheet at (13, 4), y: 4 should be between 0 and 3');
+
+    expect(ss.getSprite(14, 3)).toBeNull();
+    expect(logger.warn).toHaveBeenCalledWith('No sprite exists in the SpriteSheet at (14, 3), x: 14 should be between 0 and 13');
+
+    expect(ss.getSprite(-1, 3)).toBeNull();
+    expect(logger.warn).toHaveBeenCalledWith('No sprite exists in the SpriteSheet at (-1, 3), x: -1 should be between 0 and 13');
+
+    expect(ss.getSprite(1, -1)).toBeNull();
+    expect(logger.warn).toHaveBeenCalledWith('No sprite exists in the SpriteSheet at (1, -1), y: -1 should be between 0 and 3');
   });
 
   it('can be created from a legacy sprite sheet', async () => {